Fix failing tests.

+++ src/Traits/FrozenTimeTrait.php
@@ -41,7 +41,7 @@ trait FrozenTimeTrait
             $time = $time->format('Y-m-d 00:00:00');
         }
         if (substr($time, 0, 1) === '@') {
-            return gmdate('Y-m-d 00:00:00', substr($time, 1));
+            return gmdate('Y-m-d 00:00:00', (int)substr($time, 1));
         }
         if ($time === null || $time === 'now' || $time === '') {
             return date('Y-m-d 00:00:00');
